fix: distribute helpers into original files, remove sync_ccn_helpers.go
Codacy sees new file as +60 complexity; moving helpers back into their source files
lets Codacy compute net decrease per file (-16 sync.go, -21 sync_mempool.go, etc.)
