+++ clients/go/node/sync_mempool.go
@@ -89,3 +89,94 @@ func cloneMempoolEntry(entry *mempoolEntry) mempoolEntry {
 		source:       entry.source,
 	}
 }
+
+// buildMempoolRestoreMaps builds txid/wtxid/spender/admission maps from snapshot entries.
+func buildMempoolRestoreMaps(snapshotEntries []mempoolEntry, maxTxs int, maxBytes int) (
+	txs map[[32]byte]*mempoolEntry,
+	wtxids map[[32]byte][32]byte,
+	spenders map[consensus.Outpoint][32]byte,
+	maxAdmissionSeq uint64,
+	usedBytes int,
+	err error,
+) {
+	txs = make(map[[32]byte]*mempoolEntry, len(snapshotEntries))
+	wtxids = make(map[[32]byte][32]byte, len(snapshotEntries))
+	spenders = make(map[consensus.Outpoint][32]byte)
+	admissionSeqs := make(map[uint64][32]byte, len(snapshotEntries))
+	for _, item := range snapshotEntries {
+		entry := cloneMempoolEntry(&item)
+		if _, exists := txs[entry.txid]; exists {
+			return nil, nil, nil, 0, 0, fmt.Errorf("duplicate mempool snapshot txid %x", entry.txid)
+		}
+		if existing, exists := wtxids[entry.wtxid]; exists {
+			return nil, nil, nil, 0, 0, fmt.Errorf("duplicate mempool snapshot wtxid %x existing=%x new=%x", entry.wtxid, existing, entry.txid)
+		}
+		if err := validateMempoolSnapshotEntry(entry); err != nil {
+			return nil, nil, nil, 0, 0, err
+		}
+		if existing, exists := admissionSeqs[entry.admissionSeq]; exists {
+			return nil, nil, nil, 0, 0, fmt.Errorf("duplicate mempool snapshot admission_seq %d existing=%x new=%x", entry.admissionSeq, existing, entry.txid)
+		}
+		if len(txs) >= maxTxs {
+			return nil, nil, nil, 0, 0, fmt.Errorf("mempool snapshot exceeds transaction cap: count=%d max=%d", len(txs)+1, maxTxs)
+		}
+		if entry.size > maxBytes || usedBytes > maxBytes-entry.size {
+			return nil, nil, nil, 0, 0, fmt.Errorf("mempool snapshot exceeds byte cap: used=%d entry=%d max=%d", usedBytes, entry.size, maxBytes)
+		}
+		for _, op := range entry.inputs {
+			if existing, exists := spenders[op]; exists {
+				return nil, nil, nil, 0, 0, fmt.Errorf("duplicate mempool snapshot spender txid=%x vout=%d existing=%x new=%x", op.Txid, op.Vout, existing, entry.txid)
+			}
+			spenders[op] = entry.txid
+		}
+		entryCopy := entry
+		txs[entryCopy.txid] = &entryCopy
+		wtxids[entryCopy.wtxid] = entryCopy.txid
+		admissionSeqs[entryCopy.admissionSeq] = entryCopy.txid
+		usedBytes += entryCopy.size
+		if entryCopy.admissionSeq > maxAdmissionSeq {
+			maxAdmissionSeq = entryCopy.admissionSeq
+		}
+	}
+	return
+}
+
+// validateMempoolEntryParsed parses raw tx bytes inside a mempool entry and validates consistency.
+func validateMempoolEntryParsed(entry mempoolEntry) error {
+	tx, txid, wtxid, consumed, err := consensus.ParseTx(entry.raw)
+	if err != nil {
+		return fmt.Errorf("invalid mempool snapshot entry raw for txid %x: %w", entry.txid, err)
+	}
+	if consumed != len(entry.raw) {
+		return fmt.Errorf("mempool snapshot entry has trailing bytes for txid %x: consumed=%d raw_len=%d", entry.txid, consumed, len(entry.raw))
+	}
+	if txid != entry.txid {
+		return fmt.Errorf("mempool snapshot entry txid mismatch: entry=%x raw=%x", entry.txid, txid)
+	}
+	if wtxid != entry.wtxid {
+		return fmt.Errorf("mempool snapshot entry wtxid mismatch: entry=%x raw=%x", entry.wtxid, wtxid)
+	}
+	weight, _, _, err := consensus.TxWeightAndStats(tx)
+	if err != nil {
+		return fmt.Errorf("invalid mempool snapshot entry weight for txid %x: %w", entry.txid, err)
+	}
+	if entry.weight != weight {
+		return fmt.Errorf("mempool snapshot entry weight mismatch: entry=%d computed=%d txid=%x", entry.weight, weight, entry.txid)
+	}
+	if entry.admissionSeq == 0 {
+		return fmt.Errorf("invalid mempool snapshot entry admission_seq for txid %x: seq=0", entry.txid)
+	}
+	if !validMempoolTxSource(entry.source) {
+		return fmt.Errorf("invalid mempool snapshot entry source for txid %x: source=%q", entry.txid, entry.source)
+	}
+	if len(entry.inputs) != len(tx.Inputs) {
+		return fmt.Errorf("mempool snapshot entry input count mismatch for txid %x: entry=%d tx=%d", entry.txid, len(entry.inputs), len(tx.Inputs))
+	}
+	for i, in := range tx.Inputs {
+		want := consensus.Outpoint{Txid: in.PrevTxid, Vout: in.PrevVout}
+		if entry.inputs[i] != want {
+			return fmt.Errorf("mempool snapshot entry input mismatch for txid %x at index=%d", entry.txid, i)
+		}
+	}
+	return nil
+}
